Start moving bom logic to buildinfobom file
